moved thalicryptomanager to security/thaliCryptoManager
